+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2653](https://github.com/open-telemetry/opentelemetry-python/pull/2653))
 - Add variadic arguments to metric exporter/reader interfaces
   ([#2654](https://github.com/open-telemetry/opentelemetry-python/pull/2654))
+- Added a `opentelemetry.sdk.resources.ProcessResourceDetector` that adds the
+  'process.runtime.{name,version,description}' resource attributes when used
+  with the `opentelemetry.sdk.resources.get_aggregated_resources` API
+  ([#2660](https://github.com/open-telemetry/opentelemetry-python/pull/2660))
 - Move Metrics API behind internal package
   ([#2651](https://github.com/open-telemetry/opentelemetry-python/pull/2651))